+++ core/Patchers/AddBonesPatcher.cs
@@ -18,19 +18,20 @@ namespace WhistleWindLobotomyMod
 		// might be an easier method, but I'm too dumb to figure it out
 		// also this works so YEEEEEEEEEEEEEEEEEEAAAAAAAAA
 		// anyways, this prevents cards from dropping bones when killed by The Train ability
+		// also keeps cards from the WhiteNight event from dropping bones
 		[HarmonyPatch(typeof(ResourcesManager), nameof(ResourcesManager.AddBones))]
 		[HarmonyPostfix]
 		public static IEnumerator AddBones(IEnumerator enumerator, ResourcesManager __instance, int amount, CardSlot slot)
 		{
-			if (slot != null && slot.Card != null)
+			if (slot != null && slot.Card != null &&
+				(slot.Card.Info.GetExtendedProperty("killedByTrain") == "1" ||
+				slot.Card.Info.HasAbility(TrueSaviour.ability) || slot.Card.Info.HasAbility(Apostle.ability) || slot.Card.Info.HasAbility(Confession.ability)))
 			{
-				if (slot.Card.Info.GetExtendedProperty("killedByTrain") != "1")
-                {
-					yield return enumerator;
-					yield break;
-				}	
+				slot.Card.Info.SetExtendedProperty("killedByTrain",0);
+				amount = 0;
+				yield break;
 			}
-			amount = 0;
+			yield return enumerator;
 			yield break;
 		}
 	}