+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 ## Changelog
-### v1.0.0.0
-Initial release
-* 71 obtainable cards
-* 38 custom abilities
-* 13 special abilities
+### v1.0.2 - Prayer and Bees patch (6/28/2022)
+* Bug fixes
+  * Fixed Queen Nest ability softlocking when Queen Bee is dropped by the Mule
+* Tweaks
+  * Tweaked Confession ability to make Heretic sequence smoother
+* Balancing
+  * Cards from the WhiteNight event no longer drop bones when killed
+  * Hundreds of Good Deeds now dies if Confession is activated during a boss
+
+### v1.0.1 - Bones and Trains patch (6/27/2022)
+* Groundwork
+  * Removed the fourth zero from the in-game version number to be consistent with the Thunderstore version number
+* Bug Fixes
+  * Fixed Boons of the Bone Lord not giving bones
+  * Fixed cards not dropping bones if a copy was previously killed by The Train
+  * Fixed The Train ability being free to activate
+* Tweaks
+  * Confession ability changed to an activated-type ability
+  * Fixed an error in the README regarding The Train ability's description
+  * Can no longer activate The Train ability if there are no other cards on the board
+* Balancing
+  * Increased The Train ability activation cost 10 --> 12
+
+### v1.0.0 - Initial release (6/26/2022)
+* Additions
+  * 71 Cards
+  * 38 Abilities
+  * 13 Special abilities